chore: move ecdsa and schnorr into `signature` module
