+++ acvm/src/pwg/blackbox/signature/ecdsa.rs
@@ -17,25 +17,13 @@ use k256::{
 };
 
 use crate::{
-    pwg::witness_to_value,
     pwg::{insert_value, OpcodeResolution},
     OpcodeResolutionError,
 };
 
-fn to_u8_vec(
-    initial_witness: &WitnessMap,
-    inputs: &[FunctionInput],
-) -> Result<Vec<u8>, OpcodeResolutionError> {
-    let mut result = Vec::with_capacity(inputs.len());
-    for input in inputs {
-        let witness_value_bytes = witness_to_value(initial_witness, input.witness)?.to_be_bytes();
-        let byte = witness_value_bytes.last().unwrap();
-        result.push(*byte);
-    }
-    Ok(result)
-}
+use super::to_u8_vec;
 
-pub(super) fn secp256k1_prehashed(
+pub(crate) fn secp256k1_prehashed(
     initial_witness: &mut WitnessMap,
     public_key_x_inputs: &[FunctionInput],
     public_key_y_inputs: &[FunctionInput],
